chore: revert interface migration
