+++ src/Shared/Components/DevtronLicenseCard/index.tsx
@@ -1,3 +1,2 @@
 export { default as DevtronLicenseCard } from './DevtronLicenseCard'
 export { default as InstallationFingerprintInfo } from './InstallationFingerprintInfo'
-export type { DevtronLicenseCardProps } from './types'
