+++ src/Shared/Components/DevtronLicenseCard/types.tsx
@@ -1,22 +1,5 @@
-import { DevtronLicenseBaseDTO, LicenseStatus } from '@Shared/index'
+import { DevtronLicenseBaseDTO } from '@Shared/index'
 
 export interface InstallFingerprintInfoProps extends Pick<DevtronLicenseBaseDTO, 'fingerprint'> {
     showHelpTooltip?: boolean
 }
-
-export type DevtronLicenseCardProps = {
-    enterpriseName: string
-    expiryDate: string
-    ttl: number
-    licenseStatus: LicenseStatus
-    isTrial: boolean
-} & (
-    | {
-          licenseKey: string
-          licenseSuffix?: never
-      }
-    | {
-          licenseKey?: never
-          licenseSuffix: string
-      }
-)
