+++ src/Shared/types.ts
@@ -1073,3 +1073,20 @@ export enum LicenseStatus {
     EXPIRED = 'EXPIRED',
     REMINDER_THRESHOLD_REACHED = 'REMINDER_THRESHOLD_REACHED',
 }
+
+export type DevtronLicenseCardProps = {
+    enterpriseName: string
+    expiryDate: string
+    ttl: number
+    licenseStatus: LicenseStatus
+    isTrial: boolean
+} & (
+    | {
+          licenseKey: string
+          licenseSuffix?: never
+      }
+    | {
+          licenseKey?: never
+          licenseSuffix: string
+      }
+)
